+++ exporter/opentelemetry-exporter-otlp-proto-grpc/src/opentelemetry/exporter/otlp/proto/grpc/exporter.py
@@ -22,6 +22,8 @@ from time import sleep
 from typing import Any, Callable, Dict, Generic, List, Optional
 from typing import Sequence as TypingSequence
 from typing import Text, TypeVar
+from urllib import parse
+from urllib.parse import urlparse
 
 from backoff import expo
 from google.rpc.error_details_pb2 import RetryInfo
@@ -194,8 +196,18 @@ class OTLPExporterMixin(
         super().__init__()
 
         endpoint = endpoint or environ.get(
-            OTEL_EXPORTER_OTLP_ENDPOINT, "localhost:4317"
+            OTEL_EXPORTER_OTLP_ENDPOINT, "http://localhost:4317"
         )
+
+        parsed_url = urlparse(endpoint)
+
+        if insecure is None:
+            if parsed_url.scheme == "https":
+                insecure = False
+            else:
+                insecure = True
+
+        endpoint = parsed_url.netloc
 
         self._headers = headers or environ.get(OTEL_EXPORTER_OTLP_HEADERS)
         if isinstance(self._headers, str):