+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opentelemetry-opentracing-shim` Fix an issue in the shim where a Span was being wrapped
   in a NonRecordingSpan when it wasn't necessary.
   ([#1776](https://github.com/open-telemetry/opentelemetry-python/pull/1776))
+- OTLP Exporter now uses the scheme in the endpoint to determine whether to establish
+  a secure connection or not.
+  ([#1771](https://github.com/open-telemetry/opentelemetry-python/pull/1771))
 
 ## [1.0.0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.0.0) - 2021-03-26
 ### Added